services/weather-service/weather-service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
redis==4.5.5
psycopg2-binary==2.9.7
schedule==1.2.0
python-dotenv==1.0.0
